Fix: unload error when load plug-in fails, volume in level 2 interface now float [0,1], Optimization: filter plug-ins without private data

+++ src/pm123/plugman.h
@@ -119,7 +119,7 @@ BOOL  out_is_active( int number );
 int   out_set_active( int number );
 ULONG out_setup( const FORMAT_INFO* formatinfo, const char* URI );
 ULONG out_close( void );
-void  out_set_volume( int volume );
+void  out_set_volume( double volume ); // volume: [0,1]
 ULONG out_pause( BOOL pause );
 void  out_trashbuffers( int temp_playingpos );
 BOOL  out_flush( void );
